refactor: replace fetchPubkeyStatus with fetchBeaconStatusInChunks for improved performance and code clarity in useMyData, usePoolPubkeyData, and ValidatorSlice; standardize string quotes
