+++ hooks/useMyData.ts
@@ -2,26 +2,26 @@ import {
   getLsdEthTokenContract,
   getNetworkWithdrawContract,
   getNodeDepositContract,
-} from "config/contract";
-import { useCallback, useEffect, useMemo, useState } from "react";
-import { getEthWeb3 } from "utils/web3Utils";
-import { useWalletAccount } from "./useWalletAccount";
-import Web3 from "web3";
+} from 'config/contract';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { getEthWeb3 } from 'utils/web3Utils';
+import { useWalletAccount } from './useWalletAccount';
+import Web3 from 'web3';
 import {
   ChainPubkeyStatus,
   IpfsRewardItem,
   RewardJsonResponse,
-} from "interfaces/common";
-import { useAppSlice } from "./selector";
+} from 'interfaces/common';
+import { useAppSlice } from './selector';
 import {
   getNetworkWithdrawContractAbi,
   getNodeDepositContractAbi,
-} from "config/contractAbi";
-import { useUserPubkeys } from "./useUserPubkeys";
-import { getEthereumChainId, getValidatorTotalDepositAmount } from "config/env";
-import { formatScientificNumber, removeDecimals } from "utils/numberUtils";
-import { fetchPubkeyStatus } from "utils/apiUtils";
-import { isPubkeyStillValid } from "utils/commonUtils";
+} from 'config/contractAbi';
+import { useUserPubkeys } from './useUserPubkeys';
+import { getEthereumChainId, getValidatorTotalDepositAmount } from 'config/env';
+import { formatScientificNumber, removeDecimals } from 'utils/numberUtils';
+import { fetchBeaconStatusInChunks, fetchPubkeyStatus } from 'utils/apiUtils';
+import { isPubkeyStillValid } from 'utils/commonUtils';
 
 export function useMyData() {
   const { updateFlag } = useAppSlice();
@@ -53,13 +53,13 @@ export function useMyData() {
       }
     });
 
-    return totalManagedToken + "";
+    return totalManagedToken + '';
   }, [nodePubkeys]);
 
   const updateData = useCallback(async () => {
     if (!metaMaskAccount) {
-      setSelfDepositedToken("--");
-      setMyRewardTokenAmount("--");
+      setSelfDepositedToken('--');
+      setMyRewardTokenAmount('--');
       return;
     }
 
@@ -103,12 +103,12 @@ export function useMyData() {
       const response = await fetch(
         `https://${nodeRewardsFileCid}.ipfs.dweb.link/${getLsdEthTokenContract().toLowerCase()}-rewards-${getEthereumChainId()}-${latestMerkleRootEpoch}.json`,
         {
-          method: "GET",
+          method: 'GET',
           headers: {},
         }
       );
       const resText = await response.text();
-      var JSONbig = require("json-bigint");
+      var JSONbig = require('json-bigint');
       const resTextJson = JSONbig.parse(resText);
 
       const list: IpfsRewardItem[] = resTextJson.List?.map((item: any) => {
@@ -125,7 +125,7 @@ export function useMyData() {
       const myRewardInfo = list?.find((item) => item.address === userAddress);
       setIpfsMyRewardInfo(myRewardInfo);
 
-      const myTotalRewardAmount = myRewardInfo?.totalRewardAmount || "0";
+      const myTotalRewardAmount = myRewardInfo?.totalRewardAmount || '0';
 
       const availableExitDeposit = !myRewardInfo
         ? 0
@@ -158,7 +158,7 @@ export function useMyData() {
       const myRewardEth = Web3.utils.fromWei(
         formatScientificNumber(
           Number(myTotalRewardAmount) - Number(totalClaimedRewardOfNode)
-        ) + ""
+        ) + ''
       );
 
       setMyRewardTokenAmount(myRewardEth);
@@ -190,31 +190,26 @@ export function useMyData() {
         {}
       );
 
-      const requests = pubkeysOfNode.map((pubkeyAddress) => {
-        return (async () => {
-          const pubkeyInfo = await nodeDepositContract.methods
-            .pubkeyInfoOf(pubkeyAddress)
-            .call()
-            .catch((err: any) => {
-              console.log({ err });
-            });
+      const [pubkeyInfos, beaconStatusResponses] = await Promise.all([
+        Promise.all(
+          pubkeysOfNode.map((pubkeyAddress: string) =>
+            nodeDepositContract.methods.pubkeyInfoOf(pubkeyAddress).call()
+          )
+        ),
+        fetchBeaconStatusInChunks(pubkeysOfNode),
+      ]);
 
-          return pubkeyInfo;
-        })();
-      });
+      const beaconStatusResJson = beaconStatusResponses.flatMap(
+        (response) => response.data
+      );
 
-      const pubekyInfos = await Promise.all(requests);
       let myShareAmount = 0;
       let selfDepositAmount = 0;
 
       let totalNodeDepositAmount = 0;
 
-      const beaconStatusResJson = await fetchPubkeyStatus(
-        pubkeysOfNode.join(",")
-      );
-
-      pubekyInfos.forEach((pubkeyInfo, index) => {
-        const matchedBeaconData = beaconStatusResJson.data?.find(
+      pubkeyInfos.forEach((pubkeyInfo, index) => {
+        const matchedBeaconData = beaconStatusResJson?.find(
           (item: any) => item.validator?.pubkey === pubkeysOfNode[index]
         );
 
@@ -247,11 +242,11 @@ export function useMyData() {
       setMySharePercentage(
         Number(Web3.utils.fromWei(formatScientificNumber(myShareAmount))) /
           Number(totalManagedToken) +
-          ""
+          ''
       );
 
       setSelfDepositedToken(
-        Web3.utils.fromWei(formatScientificNumber(selfDepositAmount) + "")
+        Web3.utils.fromWei(formatScientificNumber(selfDepositAmount) + '')
       );
     } catch (err: any) {
       console.log({ err });
